Refactorings: models.py separated into python package
