+++ news/models/__init__.py
@@ -1,0 +1,9 @@
+# coding=utf-8
+import logging
+from .facebook import FacebookUser, FacebookComment, FacebookPage, FacebookPost
+from .feed import Article, Feed
+from .nltk import Tag, Word
+from .twitter import Tweet, TwitterUser
+
+__author__ = 'ilov3'
+logger = logging.getLogger(__name__)
