+++ news/models/facebook.py
@@ -1,0 +1,49 @@
+# coding=utf-8
+import logging
+
+from django.db import models
+
+from news.models.nltk import Word
+
+__author__ = 'ilov3'
+logger = logging.getLogger(__name__)
+
+
+class FacebookPage(models.Model):
+    title = models.CharField(max_length=255)
+    url = models.URLField()
+    is_active = models.BooleanField(default=False)
+
+    def __str__(self):
+        return self.title
+
+
+class FacebookPost(models.Model):
+    parent_page = models.ForeignKey(FacebookPage)
+    created_time = models.DateTimeField()
+    text = models.TextField()
+    post_id = models.CharField(max_length=255, unique=True)
+    words = models.ManyToManyField(Word)
+
+    def __str__(self):
+        return "%s..." % self.text[:40]
+
+
+class FacebookUser(models.Model):
+    user_id = models.CharField(max_length=255, unique=True)
+    name = models.CharField(max_length=255)
+
+    def __str__(self):
+        return self.name
+
+
+class FacebookComment(models.Model):
+    post_id = models.ForeignKey(FacebookPost, blank=True, null=True)
+    user_id = models.ForeignKey(FacebookUser)
+    created_time = models.DateTimeField()
+    message = models.TextField()
+    comment_id = models.CharField(max_length=255, unique=True)
+    words = models.ManyToManyField(Word)
+
+    def __str__(self):
+        return "%s..." % self.message[:40]
